% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVensemble.R
\name{CVensemble}
\alias{CVensemble}
\title{The function provides an evaluation of the performance of the algorithm. 
For this, the function separates the training data between the cross validation 
training sample (random 80% of the whole training data) and the 
cross validation test sample (the rest). The score for this one fold 
cross validation is updated and displayed after the run of each model.}
\usage{
CVensemble(data, y, rounds, threads = 2, ram = "4g")
}
\arguments{
\item{data}{An object of class data frame or matrix containing the data}

\item{y}{A character indicating the name of the response variable}

\item{rounds}{Number of iterations that should be displayed}

\item{threads}{Number of threads the H20 server should use}

\item{ram}{Maximum number of memory}
}
\value{
Prints updated accuracies of each itteration of the ensemble
}
\description{
The function provides an evaluation of the performance of the algorithm. 
For this, the function separates the training data between the cross validation 
training sample (random 80% of the whole training data) and the 
cross validation test sample (the rest). The score for this one fold 
cross validation is updated and displayed after the run of each model.
}
\examples{

# Set Dimensions of training data set
n <- 1000  
m <- 61
# Compute Create random feature matrix and response vector
features <- matrix( rnorm( n*m , 0, 10 ), nrow = n, ncol = m )
response <- sample(5 , n, replace = T)

# Create data set and rename columns
toy <- data.frame(features , response)
colnames(toy) <- c(sapply(1:m,function(x) paste0("variable",x)),"popularity")

CVensemble( data = toy, y = 'popularity', rounds = 2 )
}

