% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predForest.R
\name{predictForest}
\alias{predictForest}
\title{Performs a random forest model based on input specifications}
\usage{
predictForest(training = NULL, Nresponse = "popularity", test = NULL,
  drops = c("id", "url"), nt = 1000)
}
\arguments{
\item{training}{A data frame containing both the predicators and response}

\item{Nresponse}{Character defining the name of the response varaible in training}

\item{test}{A data frame containing the test data}

\item{drops}{A vector containing column names that should be droped from the training set}

\item{nt}{An integer specifying how many trees should be used in random}
}
\value{
A data frame containing predicted labels for each id in the \code{test} data
}
\description{
This functions reproduces in some ways the result achieved in the
in the Kaggle competition so far. The function performing based on the 
arguments a random forest model to predict the test data.
}
\examples{
# Create a simple training data set similiar to the competition set
N  <- 10
M  <- 10
# Transform to quantile matrix
predictForest( training = trSet, test =  )
}

