% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toQuantile.R
\name{toQuantile}
\alias{toQuantile}
\title{Transforms values of matrix or data frame to corresponding quantiles.}
\usage{
toQuantile(data, quant = seq(0, 1, 0.01), adj.l = 0.1^10, adj.u = 0.2^10)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{quant}{A vector between 0 and 1 with quantiles, percentiles, etc. By default set to percentiles}

\item{adj.l}{A number, which is used as the lower bound of the random noise added to the values in each column. By default is set to 0.1**10}

\item{adj.u}{A number, which is used as the upper bound of the random noise added to the values in each column. By default is set to 0.1**10 0.2**10.}
}
\value{
A matrix with transformed values of \code{data}.
}
\description{
This functions transforms each value in a column to the quantile it
it belongs to. Notice that if quantiles are overlapping the function
produces an error. To fix that there will be added a tiny random noise
(uniform distributed) to each value to avoid this circumstance.
}
\examples{
# create a matrix
mat <- matrix(rnorm(30),ncol=5, nrow= 6)

# Transform to quantile matrix
toQuantile(mat)
}

